some so-far working source splitting. This is just a start to have small source files at all (makes debugging with Windows possible).
Next thing to do is figuring out source info
and grouping pieces together.

(p.s.: Armin fixed a bug in my include file patches)

+++ pypy/translator/c/src/int.h
@@ -66,7 +66,7 @@
 #define OP_INT_MUL_OVF(x,y,r,err) \
 	{ \
 		PY_LONG_LONG lr = (PY_LONG_LONG)(x) * (PY_LONG_LONG)(y); \
-		r = lr; \
+		r = (long)lr; \
 		if ((PY_LONG_LONG)r == lr); \
 		else FAIL_OVF(err, "integer multiplication"); \
 	}
